firefox: Fix BINDGEN_CFLAGS expanding error without sysroots
It globs under sysroots for header include dir in `do_configure()`,
but it has failed if the include dir does not exist at the recipe loading
(It does not wait for `do_configure()` execution).

In this fix, delayed macro expansion until `do_configure()`,
implementing the configuration process as a Python function.